chore: moved discord bot files to a module + modified main (renamed the original main file into main.exmaple) [ALERT: THIS IS NOT WORKING]
