+++ src/main.ts
@@ -1,137 +1,92 @@
-import { Bot } from "./bot";
-import { Deploy } from "./deploy";
-import blessed from "blessed";
-import figlet from "figlet";
-import { version } from "../package.json";
+import { exit } from "process";
+import { ModuleConfigType, ModuleType } from "./types";
 
-class TUI {
-  public screen: blessed.Widgets.Screen;
+class Prime {
+  public modules: ModuleConfigType[];
+  public moduleCache: ModuleType[];
+  private threads: any[];
 
   constructor() {
-    this.screen = blessed.screen({
-      cursor: {
-        artificial: true,
-        shape: "block",
-        blink: true,
-        color: "green",
-      },
-      smartCSR: true,
-      dockBorders: true,
-      autoPadding: true,
-    });
-    this.screen.key(["escape", "q", "C-c"], function (ch, key) {
-      return process.exit(0);
-    });
+    this.modules = [];
+    this.moduleCache = [];
+    this.threads = [];
   }
 
-  public render() {
-    this.screen.render();
+  public import() {
+    if (this.modules && this.modules.length > 0) {
+      this.modules.sort((a, b) => (a!.layer > b!.layer ? 1 : -1));
+
+      this.modules.forEach((module) => {
+        const moduleClass = require(module.path).default;
+        if (moduleClass.prototype._start) {
+          this.moduleCache.push(moduleClass as ModuleType);
+        } else {
+          console.log("Module does not have a `_start` function! Skipping...");
+        }
+      });
+    } else {
+      console.log("No modules added to the Prime class!");
+      exit(1);
+    }
   }
-  public createElement(
-    element: (
-      options: blessed.Widgets.ElementOptions
-    ) => blessed.Widgets.BlessedElement,
-    properties: blessed.Widgets.ElementOptions
-  ): blessed.Widgets.BlessedElement {
-    const fullElement = element(properties);
-    this.screen.append(fullElement);
-    return fullElement;
+
+  public async start() {
+    for (const module of this.moduleCache) {
+    }
+  }
+
+  private async createThread(workerFunction: () => any) {
+    // Convert the function to a string
+    const functionCode = workerFunction.toString();
+
+    // Create a Blob with the worker code
+    const blob = new Blob(
+      [
+        `
+            self.onmessage = () => {
+                (${functionCode})();
+                self.postMessage("Task has finished execution");
+            };
+        `,
+      ],
+      { type: "application/javascript" }
+    );
+
+    // Create a URL for the Blob
+    const workerURL = URL.createObjectURL(blob);
+
+    // Create the Web Worker
+    const worker = new Worker(workerURL);
+
+    // Send a start message to the worker
+    worker.postMessage("start");
+
+    // Handle messages from the worker
+    worker.onmessage = (event) => {
+      console.log(event.data);
+    };
+
+    // Handle errors
+    worker.onerror = (error) => {
+      console.error("Error in worker:", error.message);
+    };
+
+    return worker;
   }
 }
 
 async function main() {
-  const screen = new TUI();
+  const prime = new Prime();
 
-  const loadingLog = screen.createElement(blessed.log, {
-    top: "50%",
-    left: "20%",
-    width: "40%",
-    height: "50%",
-    label: "Loading Log",
-    content: "",
-    tags: true,
-    border: {
-      type: "line",
+  prime.modules = [
+    {
+      path: "./modules/discordbot/bot.js",
+      layer: 1,
     },
-    style: {
-      fg: "green",
-      border: {
-        fg: "white",
-      },
-    },
-    scrollable: true,
-  });
+  ];
 
-  const errorLog = screen.createElement(blessed.log, {
-    top: "50%",
-    right: 0,
-    width: "40%",
-    height: "50%",
-    label: "Error Log",
-    content: "",
-    tags: true,
-    border: {
-      type: "line",
-    },
-    style: {
-      fg: "red",
-      border: {
-        fg: "white",
-      },
-    },
-  });
-
-  const loadedCog = screen.createElement(blessed.box, {
-    top: "50%",
-    left: 0,
-    width: "20%",
-    height: "50%",
-    label: "Loaded Cogs",
-    content: "- Loading...",
-    tags: true,
-    border: {
-      type: "line",
-    },
-    style: {
-      fg: "blue",
-      border: {
-        fg: "white",
-      },
-    },
-  });
-
-  const title = screen.createElement(blessed.box, {
-    top: 0,
-    left: 0,
-    width: "100%",
-    height: "50%",
-    align: "center",
-    valign: "middle",
-    content:
-      figlet.textSync("DisCogs", {
-        font: "Small Shadow",
-        horizontalLayout: "default",
-        verticalLayout: "default",
-      }) + `\nv${version}`,
-    tags: true,
-    border: {
-      type: "line",
-    },
-    style: {
-      border: {
-        fg: "white",
-      },
-      fg: "red",
-    },
-  });
-
-  const deploy = new Deploy(loadingLog, errorLog, loadedCog);
-  await deploy.registerCommands();
-
-  const bot = new Bot(loadingLog, errorLog, loadedCog);
-  await bot.start();
+  prime.import();
+  console.log(prime.moduleCache);
 }
 
-main().catch((error) => {
-  console.error("An error occurred while starting the bot:", error);
-});
+main();
