+++ src/modules/discordbot/cogs.ts
@@ -1,4 +1,4 @@
-import { LegacyCommand, SlashCommand, EventFunction } from "./types";
+import { LegacyCommand, SlashCommand, EventFunction } from "../../types";
 
 export class Cog {
   name: string;
