+++ src/modules/discordbot/cogs/ping.ts
