+++ src/modules/discordbot/util/log.ts
